==> 5- Real Life Performance.ipynb <==
In this notebook, I will be testing my algoirthm(s) on unseen data. As opposed to
cross validation


import time

# import libraries
from surprise import Dataset
from surprise.accuracy import rmse
from own_algorithms.UserItemKNN import UserItemKNN
from surprise import KNNBasic
from own_algorithms.UserItemKNNv2 import UserItemKNNv2
from surprise.model_selection import train_test_split
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import random
from own_algorithms.top_n_list import get_top_n_list


# load data and split into A/B sets, A for CFV and B for unbiased testing
# Load the data using the built-in function
data = Dataset.load_builtin('ml-100k')
raw_ratings= data.raw_ratings

random.seed(2001)
np.random.seed(2001)
random.shuffle(raw_ratings)

# create threshold for unseen, 80-20
cutoff = int(0.8 * len(raw_ratings))
A_raw= raw_ratings[:cutoff]
B_raw= raw_ratings[cutoff:]

# data is now only set A ratings
data.raw_ratings= A_raw

# testing the first version ensemble
algo= UserItemKNN()

trainset = data.build_full_trainset()
algo.fit(trainset)


testset = data.construct_testset(B_raw)  # testset is now the set B
predictions = algo.test(testset)
print("Unbiased accuracy on B for v1,", end=" ")
algo1_rmse= rmse(predictions)

#testing the "improved version"
algo2= UserItemKNNv2()
algo2.fit(trainset)
predictions= algo2.test(testset)
print("Unbiased accuracy on B for v2,", end=" ")
algo2_rmse= rmse(predictions)

algo3= KNNBasic()
algo3.fit(trainset)
predictions= algo2.test(testset)
print("Unbiased accuracy on B for KNNBasic,", end=" ")
algo3_rmse= rmse(predictions)

These scores are for the 100k dataset. This is not a realistic dataset as each film,user pair is filled.
Most real life predictions are made using a sparse dataset. The next cells will be using the ml-1m dataset.
A larger movie rating dataset with approximatly 95% sparsity. It is expected the performance will drop


#load the 1m dataset
data=Dataset.load_builtin('ml-1m')

#the algos havent been optimisded on data so its unncesccasy to split the dataset to validaton and test sets
trainset, testset = train_test_split(data, test_size=0.25, random_state=1)
# Train the algorithm on the trainset, and predict ratings for the testset

algo.fit(trainset)
predictions = algo.test(testset)
print("Unbiased accuracy on B for v1,", end=" ")
algo1_rmse_1m= rmse(predictions)

fit_start=time.time()
algo2.fit(trainset)
fit_hybrid=time.time()-fit_start
predict_start=time.time()
predictions= algo2.test(testset)
predict_hybrid=time.time()-predict_start
print("Unbiased accuracy on B for v2,", end=" ")
algo2_rmse_1m= rmse(predictions)

hybrid_stats= np.array([algo2_rmse_1m, fit_hybrid, predict_hybrid])

hybrid_stats


movies_cols = ['movie_id', 'title', 'genres']
movies_df = pd.read_csv('./ml-1m/movies.dat', sep='::', names=movies_cols, engine='python', encoding='latin-1')

# create top n list
movies=get_top_n_list(predictions, 10, '398', movies_df)
df398= pd.DataFrame({'Hybrid':movies})
movies=get_top_n_list(predictions, 10, '1', movies_df)
df1= pd.DataFrame({'Hybrid':movies})
movies=get_top_n_list(predictions, 10, '134', movies_df)
df134= pd.DataFrame({'Hybrid':movies})

df398

df1

df134

Well, the results here show us that our algorithm actually performs better in the ML-1M dataset.
THis was an unexpected result however the tests do not lie. This could be due to the fact the 1m dataset has 10x
the ratings than the 100k and theefore can make more accuracte predictions. I will test the KNNBasic algorithm again to compare the results



algo3=KNNBasic()
fit_start=time.time()
algo3.fit(trainset)
fit_basic= time.time()- fit_start
predict_start=time.time()
predictions= algo3.test(testset)
predict_basic=time.time()-predict_start
print("Unbiased accuracy on B for KNNBasic,", end=" ")
algo3_rmse_1m= rmse(predictions)

basic_stats=np.array([algo3_rmse_1m, fit_basic, predict_basic])
basic_stats


data=pd.DataFrame(columns=['RMSE', 'Fit Time', 'Predict Time'])
data.loc[len(data)] = basic_stats
data.loc[len(data)] = hybrid_stats
data.insert(0,'Algorithm', ['KNN', 'KNN (hybrid)'])
data.to_csv('./algo_data/KNN_1m', index=False)

movies=get_top_n_list(predictions, 10, '398', movies_df)
df398["KNN Basic"]= movies
movies=get_top_n_list(predictions, 10, '1', movies_df)
df1["KNN Basic"]= movies
movies=get_top_n_list(predictions, 10, '134', movies_df)
df134["KNN Basic"]= movies


df398

df1

df134

df398.to_csv('./predictions/398.csv', index=False)
df1.to_csv('./predictions/1.csv', index=False)
df134.to_csv('./predictions/134.csv', index=False)

data={"Algo":["KNNBasic", "V1", "V2"],
      "100k": [algo3_rmse, algo1_rmse, algo2_rmse],
      "1M": [algo3_rmse_1m, algo1_rmse_1m, algo2_rmse_1m]}

results= pd.DataFrame(data)
results.to_csv('./algo_data/KNN_100Kvs1M.csv', index=False)

results
