==> 6- Sparsity.ipynb <==
In this notebook, the KNNBaseline algorithm will be introduced. User bias can be an issue
in certain datasets and using the baseline aims to improve predictions by incorporating a baseline rating


import time

# import libraries
from surprise import Dataset
from surprise.accuracy import rmse
from own_algorithms.UserItemKNN import UserItemKNN
from surprise import KNNBaseline
from own_algorithms.UserItemKNNv2 import UserItemKNNv2
from surprise.model_selection import train_test_split
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import random
from surprise.model_selection import GridSearchCV



Before running the algorithm, we are gonna test for the best parameter setup. Unlike earlier, GridSearch will be used for sake of simplicity.

# takes a while to run- 21 mins
data= Dataset.load_builtin('ml-100k')
param_grid = {'k': [10, 20, 30, 40, 50, 60],
              'sim_options': {'name': ['cosine', 'pearson_baseline'],
                              'min_support': [1, 5],
                              'user_based': [True, False]}}


# Instantiate the GridSearchCV object and fit the data
gs = GridSearchCV(KNNBaseline, param_grid, measures=['rmse', 'mae'], cv=5)
gs.fit(data)

# Print the best RMSE score and the corresponding parameters
print(gs.best_score['rmse'])
print(gs.best_params['rmse'])

algo = gs.best_estimator['rmse']

#test on 100k
trainset, testset = train_test_split(data, test_size=0.25)

algo.fit(trainset)

predictions = algo.test(testset)
print("Unbiased accuracy for KNNBaseline on ml100k,", end=" ")
algo_rmse= rmse(predictions)

#test on 1m
data=Dataset.load_builtin('ml-1m')
trainset, testset = train_test_split(data, test_size=0.25)
fit_start= time.time()
algo.fit(trainset)
fit_baseline= time.time()-fit_start
predict_start=time.time()
predictions = algo.test(testset)
predict_baseline=time.time()-predict_start
print("Unbiased accuracy for KNNBaseline on ml-1m,", end=" ")
algo_rmse_1m= rmse(predictions)
baseline_stats= np.array([algo_rmse_1m, fit_baseline, predict_baseline])

KNNbaselineScores= pd.DataFrame(data={"Algo":["KNNBaseline"],
                                "100k": [algo_rmse],
                                "1M":[algo_rmse_1m]})
KNNbaselineScores.to_csv('./algo_data/KNNBaselineOp.csv', index=False)

baseline_stats

# load knn scores and update with
data= pd.read_csv('./algo_data/KNN_1m')
new_row= {'Algorithm': 'KNN Baseline', 'RMSE': algo_rmse_1m, 'Fit Time': fit_baseline, 'Predict Time': predict_baseline}
data.loc[len(data)] = new_row
data.to_csv('./algo_data/KNN_1m', index=False)

movies_cols = ['movie_id', 'title', 'genres']
movies_df = pd.read_csv('./ml-1m/movies.dat', sep='::', names=movies_cols, engine='python', encoding='latin-1')

# load dataframes with predictions
user1= pd.read_csv('./predictions/1.csv')
user134= pd.read_csv('./predictions/134.csv')
user398= pd.read_csv('./predictions/398.csv')


from own_algorithms.top_n_list import get_top_n_list

movies=get_top_n_list(predictions, 10 ,'398', movies_df)
user398["KNN Baseline"]= movies
user398.to_csv('./predictions/398.csv', index=False)

movies=get_top_n_list(predictions, 10 ,'1', movies_df)
user1["KNN Baseline"]= movies
user1.to_csv('./predictions/1.csv',index=False)

movies=get_top_n_list(predictions, 10 ,'134', movies_df)
user134["KNN Baseline"]= movies
user134.to_csv('./predictions/134.csv', index=False)

user134

user1
