==> 7- KNN Final Analysis.ipynb <==
As the final part of the KNN analysis. It is time to compare our three algorithms, KNNBasic, My own hybrid
and KNNBaseline.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

df1= pd.read_csv("./algo_data/KNN_100Kvs1M.csv")
df2= pd.read_csv("./algo_data/KNNBaselineOp.csv")


results = pd.concat([df1, df2], ignore_index=True)
results= results.drop(1)
results=results.reset_index().drop('index',axis=1)
print(results)

knn= results.loc[0].values[1:3]
hybrid= results.loc[1].values[1:3]
KNNbase=results.loc[2].values[1:3]
